requirements.txt
requests~=2.26.0
aiogram~=2.17.1
aiogram_dialog ~=1.3.0
python-dotenv~=0.19.2
SQLAlchemy~=1.4.31